refactor(fbuild-build): split stm32/orchestrator.rs into submodules to satisfy LOC gate
